fastapi/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
scikit-learn==1.6.1
joblib==1.3.2
numpy==1.26.0
python-dotenv==1.0.0
pandas==2.0.3
matplotlib==3.7.1
seaborn==0.12.2
